Fix bugs

+++ sunpy/timeseries/sources/norh.py
@@ -114,7 +114,7 @@ class NoRHTimeSeries(GenericTimeSeries):
         obs_start_time=parse_time(header['DATE-OBS'] + 'T' + header['CRVAL1'])
         length = len(data)
         cadence = np.float(header['CDELT1'])
-        sec_array = np.linspace(0, length-1, (length/cadence))
+        sec_array = np.linspace(0, length-1, int(length/cadence))
 
         norh_time = []
         for s in sec_array:
